Apply collapsible behaviour on all fragments
