+++ src/components/DiagramFrame/SeqDiagram/MessageLayer/Block/Statement/Fragment/FragmentOpt.vue
@@ -2,13 +2,16 @@
   <div class="fragment opt border-skin-fragment rounded" :style="fragmentStyle">
     <comment v-if="comment" :comment="comment" />
     <div class="header bg-skin-fragment-header text-skin-fragment-header text-base leading-4">
-      <div class="name font-semibold p-1 border-b"><label>Opt</label></div>
+      <div class="name font-semibold p-1 border-b">
+        <collapse-button label="Opt" @click="this.toggle"/>
+      </div>
     </div>
-    <block
-      :style="{ paddingLeft: `${offsetX}px` }"
-      :context="opt.braceBlock().block()"
-      :selfCallIndent="selfCallIndent"
-    ></block>
+    <div :class="{hidden: collapsed}">
+      <block
+        :style="{ paddingLeft: `${offsetX}px` }"
+        :context="opt.braceBlock().block()"
+        :selfCallIndent="selfCallIndent"></block>
+    </div>
   </div>
 </template>
