+++ src/components/DiagramFrame/SeqDiagram/MessageLayer/Block/Statement/Fragment/FragmentTryCatchFinally.vue
@@ -2,45 +2,48 @@
   <div class="fragment tcf border-skin-fragment rounded" :style="fragmentStyle">
     <div class="segment">
       <comment v-if="comment" :comment="comment" />
-      <div
-        class="header bg-skin-fragment-header text-skin-fragment-header text-base leading-4 rounded-t"
-      >
-        <div class="name font-semibold p-1 border-b"><label>Try</label></div>
+      <div class="header bg-skin-fragment-header text-skin-fragment-header text-base leading-4 rounded-t" >
+        <div class="name font-semibold p-1 border-b">
+          <collapse-button label="Try" @click="this.toggle"/>
+        </div>
       </div>
-      <!-- fragment-offset set as offsetX - 1 for fragment border     -->
-      <block
-        v-if="blockInTryBlock"
-        :style="{ paddingLeft: `${offsetX}px` }"
-        :context="blockInTryBlock"
-        :selfCallIndent="selfCallIndent"
-      ></block>
     </div>
-    <template v-for="(catchBlock, index) in tcf.catchBlock()" :key="index + 500">
-      <div class="segment mt-2 border-t border-solid">
-        <div class="header text-skin-fragment" :key="index + 1000">
-          <label class="keyword catch p-1">catch</label
-          ><label class="exception p-1">{{ exception(catchBlock) }}</label>
-        </div>
+    <div :class="{hidden: collapsed}">
+      <div class="segment">
+        <!-- fragment-offset set as offsetX - 1 for fragment border     -->
         <block
+          v-if="blockInTryBlock"
           :style="{ paddingLeft: `${offsetX}px` }"
-          :context="blockInCatchBlock(catchBlock)"
-          :selfCallIndent="selfCallIndent"
-          :key="index + 2000"
-        ></block>
+          :context="blockInTryBlock"
+          :selfCallIndent="selfCallIndent" ></block>
       </div>
-    </template>
-    <template v-if="finallyBlock">
-      <div class="segment mt-2 border-t border-solid">
-        <div class="header text-skin-fragment finally">
-          <label class="keyword finally p-1">finally</label>
+      <template v-for="(catchBlock, index) in tcf.catchBlock()" :key="index + 500">
+        <div class="segment mt-2 border-t border-solid">
+          <div class="header text-skin-fragment" :key="index + 1000">
+            <label class="keyword catch p-1">catch</label
+            ><label class="exception p-1">{{ exception(catchBlock) }}</label>
+          </div>
+          <block
+            :style="{ paddingLeft: `${offsetX}px` }"
+            :context="blockInCatchBlock(catchBlock)"
+            :selfCallIndent="selfCallIndent"
+            :key="index + 2000"
+          ></block>
         </div>
-        <block
-          :style="{ paddingLeft: `${offsetX}px` }"
-          :context="finallyBlock"
-          :selfCallIndent="selfCallIndent"
-        ></block>
-      </div>
-    </template>
+      </template>
+      <template v-if="finallyBlock">
+        <div class="segment mt-2 border-t border-solid">
+          <div class="header text-skin-fragment finally">
+            <label class="keyword finally p-1">finally</label>
+          </div>
+          <block
+            :style="{ paddingLeft: `${offsetX}px` }"
+            :context="finallyBlock"
+            :selfCallIndent="selfCallIndent"
+          ></block>
+        </div>
+      </template>
+    </div>
   </div>
 </template>
